+++ sdk/spring/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 - GA the `spring-cloud-azure-starter-storage`. This starter supports all features of Azure Storage.
+- GA the `spring-cloud-azure-starter-keyvault`. This starter supports all features of Azure Key Vault.
 
 ### Breaking Changes
 
@@ -17,6 +18,8 @@ This section includes changes in `spring-cloud-azure-autoconfigure` module.
 #### Breaking Changes
 + Add `AzureStorageConfiguration` to make Azure storage service share common property configuration [#29094](https://github.com/Azure/azure-sdk-for-java/pull/29094).
     +  Add properties `spring.cloud.azure.storage.endpoint`, `spring.cloud.azure.storage.account-key`, `spring.cloud.azure.storage.sas-token`, `spring.cloud.azure.storage.connection-string`, `spring.cloud.azure.storage.account-name`.
++ Add `AzureKeyVaultConfiguration` to make Azure Key Vault service share common property configuration [#29163](https://github.com/Azure/azure-sdk-for-java/pull/29163).
+    + Add properties `spring.cloud.azure.keyvault`.
 
 ### Spring Messaging Azure Service Bus
 This section includes changes in the `spring-messaging-azure-servicebus` module.